requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.3
django-request==1.6.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
flake8==7.1.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
mccabe==0.7.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
